refactor(client-3d): extract ChatBubble3D + Nametag from PlayerEntity
Phase 3B: split PlayerEntity.tsx 898→378 LOC by moving the 3D chat
bubble system (SingleBubble, ChatBubble, texture cache, geometry
helpers) and the Nametag component into their own files. Jump physics
stay in PlayerEntity for now — tightly coupled to the main component's
refs and useFrame loop.
